ShellCheck: Workaround: Skip ShellCheck if executable found
Cabal should skip ShellCheck if it is already installed:

```
$ cabal install ShellCheck
Resolving dependencies...
All the requested packages are already installed:
ShellCheck-0.4.6
Use --reinstall if you want to reinstall anyway.
```

but it just doesn't work in Travis CI build environment.  For now we simply skip the installation if we found the built executable.

Signed-off-by: 林博仁 <Buo.Ren.Lin@gmail.com>